requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
MarkupSafe==1.1.1
gunicorn==20.1.0
svgwrite==1.4.1
google-cloud-storage==2.3.0
flask-dance==3.2.0
flask-login==0.6.2
